Always exit 0 or 1 based on test result
A regression was causing tap to exit with a code of 0, even
if there were failing tests.

This is a huge problem!  It means that failures were not getting
reported properly in some cases.  Thankfully, the tap runner was
exiting non-zero when there were failures, but directly running
the test script would not exit properly on failure.