test: Emit end much more immediately
This introduces an interesting and, in my opinion, preferable change in
semantics, by making the actions of the Test class even more
synchronous and directly tied to the TAP output that is generated.

The most interesting edge case that changes is when a throw occurs after
a test has ended, either via an explicit t.end(), or by fulfilling its
plan.

For example:

    var t = require('tap')
    t.test('child test', function (tt) {
      tt.plan(1)
      tt.pass('this immediately ends the child test')
      console.log('# about to throw')
      throw new Error('boom')
    })
    console.log('# after child')

Previously, the output from this would be like this: (diags elided)

    TAP version 13
        # Subtest: child test
        1..1
        ok 1 - this immediately ends the child test
    # about to throw
    # after child
    not ok 1 - child test # time=4.435ms
    not ok 2 - Error: boom
    1..2
    # failed 2 of 2 tests

That is, the child would be able to tell that the error came through,
but only at the expense of delaying its 'end' event.  This delay causes
unfortunate and confusing interactions with console.log() calls,
virtually guaranteeing that they'll always be out of order.  Note that
the actual child 'end' event (and corresponding TAP output) comes
*after* the "after child" log.

With this change, it now produces this output:

    TAP version 13
        # Subtest: child test
        1..1
        ok 1 - this immediately ends the child test
    ok 1 - child test # time=4.453ms
    # about to throw
    not ok 2 - Error: boom
    # after child
    1..2
    # failed 1 of 2 tests

So, on the one hand, we don't know that the child test threw, since its
'end' event is already fully processed by the time the throw happens.

However, the stack and other diagnostic information will still track the
throw directly to its source.  Furthermore, we always had to deal with
throws coming after a test ends, anyway.

To mitigate this effect, a thrown Error object will have the name of its
originating Test context added to it as the 'test' field.