verify that top-level tests can be defined asynchronously
This was an issue wiht TAP 0.x which I was reminded of recently.

Since the rewrite, it's not a problem, because the underlying code
was completely obliterated and designed in a cleaner way.  However,
it's always good to keep the ratchets around, so this test is here
to prevent back-sliding on this issue.