Integrate coverage, automagic coveralls.io support
Now, if you have a COVERALLS_REPO_TOKEN in the environment, it'll
automatically run tests with coverage turned on, and pipe the results to
coveralls.

So, if you're using tap for tests, and Travis-CI for CI, and you put a
COVERALLS_REPO_TOKEN in your Travis-CI envionment, then it'll all Just
Work.

Huge thanks to @BenjaminCoe for pushing on making nyc work.  Without it,
there would be no reasonable way to get coverage for CLI apps, or to
provide coverage in tap, since it runs all tests in child processes.