Handle expected non-Error objects from t.throws
Fixes #62

Comes with the unfortunate caveat that any 'extra' data passed to
t.throws (ie, '{skip:true}' and the like) requires that an expected
error be passed in.

While this is technically a breaking change, it's a very rare edge case.
In practice, almost no one ever uses (or knows about!) the diag object,
so it's probably fine.  If there's complaining, I'll revert in the next
patch and push a major bump.