support doing coverage even in node_modules
The --cov argument was looking at global.__coverage__, which is fine
in tests of this module, or when linked, but when installed as a dep,
and run via an npm test script, it would fail to realize that coverage
was loaded.

Specify in a cli flag to the child process.