Handle exit codes in a way that works for older nodes
Prior to Node 0.12, setting process.exitCode did not work.  A module
called 'exit-code' was designed to handle this, but it has problems
with the way that 'signal-exit' monkey-patches process.reallyExit,
becaus signal-exit sets process.exitCode as well.

So, the solution here is to ditch using the exit-code module, and
just set the code using the old process.exit(code) method.