Add t.tearDown(function), t.autoend()
This one is a bit of a sticky API puzzle.

We run tests synchronously, and the root tap object is automatically
ended when the process exits.

However, in a situation like this, the process will never exit:

    tap.tearDown(function () {
      server.close()
    })

because the server will keep it open.

To avoid this problem, what we do is automatically end when there is
nothing in the queue.  The autoend is scheduled in a timeout that in
turn sets another timeout.  This gives us a bit of a chance to add more
tests in this tick (or the very next one) but any delay-added tests once
an autoexit is triggered will be lost.

So, in a situation like this:

    tap.tearDown(function () {
      console.log('ended...')
    })

    tap.pass('this is ok')

    setTimeout(function () {
      tap.pass('this is lost')
    }, 100)

the `'this is lost'` test point is never printed, because we've already
hit the `autoend` which triggered the teardown.

It seems relatively unlikely to have both slowly asynchronously added
tests, *and also* a tearDown function, and any expectation that that
could ever work, since it would require knowledge of the future, and
most tap users have watched a few time travel movies and are aware that
this module does not change the laws of causality.

Nevertheless, we may need to figure out a good solution to this if it
trips people up.  The compromises in this commit make it pass tests for
`st` which uses it in a very conventional way.

Thanks to @aredridel for rubber-ducking this in the #npm irc channel.