Fail on unfinished tests
Fix #161
Fix #152
Re #129

This is a pretty significant change.  It may introduce some annoying
test failures in some cases, but most likely, those are tests that were
silently being ignored before.

Now, when `t.endAll()` is called, if there's anything waiting in the
queue, it will be treated as a failure, and reported with extreme
prejudice.  It should be virtually impossible after this change to ever
have a test simply be ignored or have the tap stream cut off after
starting a new child test.