Remove domains
Fixes a problem where expected EventEmitter-related throws are
impossible to test, because the domain hijacks the 'error' event.

Of course, this means that asynchronous unexpected throws will always be
assigned to the global root TAP object, but... what can you do?  If
people complain, we can try to figure out something clever using
continuation-local-storage or something.

Ugh.  Domains.  The not-quite-worst-possible solution to any problem,
which always ends up regrettably.

I am, as always, truly and deeply sorry for inflicting this feature on
Node.js