Add fast smoke test method
This refactors the test/test.js file that it can be easily piped to tap
and used to print out a fast smoke-test to use in dev.

Some of the other tests, especially timeout things, get pretty slow at
times.  This is also a much more satisfying way to see what it's doing.