Wait for returned promises before ending tests
Fixes #240.

If a promise is returned then that promise should dictate the test lifecycle.
This prevents tests from ending prematurely when a planned assertion count is
reached, but further work takes place before the returned promise is fulfilled
(or indeed rejected).

This commit defers ending the test until the test function has returned. Care is
taken to ensure that ending tests more that once, exceeding the planned assertion
count, or causing exceptions after ending the test, is still reported correctly
*after* the initially deferred end has finished.

A new "promise-plan" test has been added for the scenario outlined in #240.