Rewrite main parsing function in modular chunks
Fix #25
Fix #23

This moves around some things to make the following changes:

1. Stricter ordering of events.  Now a bailout will push a pending
assert event, rather than omitting it.

2. `Subtest` comments are now no longer semantically relevant.  They`re
just comments.  What makes a child test is having indented non-comment
TAP data.

3. Support for buffered subtests is re-added, and made a bit more
robust.

4. Invalid TAP that causes the parser to end in failure now also adds an
entry in the `failures` array, so it's possible to see what went wrong.

This is a breaking change, because of the semantic changes, but it
should be an easy upgrade for node-tap, its primary consumer.  There's a
possibility that it'll break some of the reporting, though, so it's
still very provisional.