Treat empty tests the same as a 'skip all' 1..0
Node-tap jumps through a lot of hoops to treat empty subtest processes
as an empty test, by synthetically adding a `1..0` line to the stream if
no valid TAP output is encountered.

Why not just do that in the parser?  It simplifies a lot of stuff.

I doubt that anyone is depending on this being a failure, but just in
case, it should be a major version bump.