split up output tests by type, not name
Juggling the alphabet is too hard as the number of tests and cases
increases.  Passing two booleans to a function feels like it should
probably be an options object, but whatever.  We can cross that bridge
later.

Also, this allows running a single case in an ad-hoc manner which is a
bit more convenient in development.