output tests for consuming static child test output
Simulates cases where the 'TAP_etc' environs aren't handled in any sort
of way in the child process test.

Particularly, the output for bail-on-failure and buffered parent when
the non-buffered non-bail-on-fail child has a failure midstream, is not
very ideal.  Printing the extra data is debatably valuable, but the fact
that it lists the test name as the bailout reason rather than the original
failure is very unfortunate.

Ideally, it'd be treated just like when a non-child-process test bails out.