Do not ignore version lines that are in yaml
This also has the effect of moving the 'line' event that starts a child
to *after* the child event that it starts, so that we can parse lines
coming out of the child rather than emitting 'line' on the way in.

The downside is that we have to be careful to emit a 'line' event once
we know what the thing is (ie, not a version, not something for a child,
not something that starts a child), so `emit('line')` function calls
happen in many more places.

But the ordering is somewhat more reasonable, in a way.  The child
object is created, and then we see its lines coming out.  And, it's
now possible to use the 'line' events as a pass-through stream that
filters out nested `TAP version 13` lines properly, cuts out
post-bailout noise, and so on.