Ensure that process.on('exit') handler is run
This MAY fix #345, but without knowing the root cause of that problem,
it's hard to say for sure.

What IS definitely fixed by this is the behavior when a test program
does this:

```js
process.on('exit', function () {
  process.exit()
})
```

prior to TAP loading and adding its exit event handler.

This also prevents TAP from hooking up to the process.on('exit') event
if it's not piping anywhere, so just doing this won't change your exit
code:

```js
var Test = require('tap').Test
```