Fix teardown/startup timing issues
DRY the 'teardown' event emission, and make the timing deterministic in
the process.

Previously, the 'teardown' event was emitted after the parent's call to
printResult.  Since printResult calls process(), which can in turn start
another subtest in motion, this caused problems because the next test's
initialization function would fire _before_ the 'teardown' event.

This change shifts a method onto the queue that will emit the teardown
event for the child, so when the call to printResult calls 'process()',
that makes the right thing happen.

Fix #353