Don't require argv file list if specified in taprc
Previously, checking for specified test files was done by looking at
argv.length.  So, if a `files` config was specified in taprc, it would
work if a `--` was passed in on the command line, but otherwise would
dump the usage message one sees when running `tap` with no arguments (or
would wait for stdin if not run on a TTY).

Fix #438