Make child nodes handle key printing/indentation
This makes the child node fully responsible for the line, including the
key portion.  Previously, the initial indentation of a key (or value in
the case of Arrays and Sets) would be done by the parent.  However, for
complex keys and values, the nested members would still need to be
managed by the child, so it got rather split.  Additionally, when using
this for generating object diffs, it meant that getting the + and - in
the right places was pretty complicated.

Also, add reference IDs for circular object handling, since it always
bugs me so much to have to go digging for that.