Handle weird circular obj edge case
If the 'expect' object contains a circular link to the reference object,
then it can create a situation where it is a simple match, but
improperly circular.  Also, if the reference object isn't circular at a
given point, but the expected object is, and the objects are identical
references, then this would confuse things.

Always check for a circular expectation, and let circular mismatches
trump simple object identity matches.