Emit 'extra' right away, so order is preserved
Waiting for a current test to end has the result of making non-tap data
be squished together and re-sorted in odd ways.  To fully accomplish
this, proxy all 'extra' events up to the parent, indenting along the
way.

This fixes https://github.com/tapjs/node-tap/issues/511 but it is a
major version change.