Move the onbail->endAll() logic into Test
This makes the Tap object less special, and easier to mock.  Also, more
of the right kinds of magic are in the right place.  The root Tap object
should be special by virtue of being a harness that mucks with process
and stdio, not by virtue of being a root Test object.

Any special behavior that is inherent to being a root Test object should
be handled by the Test class, conditional upon having no parent.

Ran into this while developing treport, and running into some
hair-pulling situations trying to get it to produce bailout reports
properly.  Without the endAll() hook, the root Test object just kinda
falls off a cliff and the output ends weirdly.

An interesting side effect of this is that we can no longer ever have a
situation where an indented subtest is started and never ended.  But, it
does open up a new case where that indented subtest is ended, but is not
the "occupied" current test.  That's why some assertions and
conditionals moved around in other parts of the class.