Add t.expectUncaughtException(wanted)
Fix #475

This is much easier to accomplish now that tap is using
async-hooks-domain instead of old-style domains for error tracking
across async scopes.

Multiple uncaught exceptions can be handled, but they must be thrown in
the precise order expected.  If a test ends (implicitly or otherwise)
with expected uncaught exceptions not having been seen, then that is a
failure.