Do not attempt to write snapshot file in runner
The runner frequently has a very long snapshot filename, since it
includes all the CLI arguments, and people may run tap like
`tap lib/**/*.test.js` and pull in a lot of files for a large program.

Since the runner never has a snapshot to save, setting it in
writeSnapshot mode means that it will try to rm the file, resulting in
an ENAMETOOLONG error.

Close #567