Load snapshot object lazily
When a snapshot is create using --node-arg configs, this can add things
that are stripped out by NYC's spawn-wrap, but this results in a
snapshot file that includes "node" in the name.

Since it doesn't happen when run without coverage, it means that a
coverage-enabled test uses a different snapshot file than the
non-coverage-enabled test run created.

Also, this means that we never create Snapshot objects that don't end up
being used (for example, for the TAP object used by the test runner).

Note that this does require capturing the cwd at module load time, or
else the Snapshot file location may change before it's used (perhaps to
a directory that has been cleaned up by a t.teardown() method).

Fix #571