Properly handle test aborts when awaiting promises
Fix #583

Add an 'abort' method to the Waiter class in lib/test.js.  This lets us
effectively bail out of a promise when `t.endAll()` is called.

Previously, using bailouts in the presence of t.resolves, t.rejects,
t.resolveMatch, or t.resolveMatchSnapshot would throw a TypeError,
because the endAll method expected this.occupied to be a Test object.

However, as of 63f9c92effe9b50f23a671ee8dd3c1ad4bb41f3a, `this.occupied`
is set to a Waiter object instead, to avoid an extra layer of
indentation for promise-handling assertions.