Fail if error is thrown while awaiting a Promise
This was causing a test in make-fetch-happen to hang silently and time
out rather than failing appropriately.

Note that this can only happen if the error is thrown _outside_ of the
Promise flow of control.  Throws from withing .then()/.catch() methods
will reject the promise, which may be what you want.  If an EventEmitter
or something throws an error, tap will still track it, but the Promise
chain might not.  If it makes it to tap, then the Promise chain is
probably broken.