Use jsdiff for all diffs
jsdiff does a better job cleaning up and fragmenting diffs, especially
in cases like [1,2,3,4,5] compared with [2,3,4,5,6].

This implementation is pretty inefficient.  Next step is to refactor it
so that Same and subclasses build up two strings to be diffed, instead
of a single patch string which is then split apart and reconstructed for
jsdiff.