Refactor to rely on jsdiff more efficiently
While it would be nice to not bother creating a diff just to examine it
and create a diff in a different way, that will require much more
elaborate re-thinking of how things are done in this module.

At least this approach is more efficient, and avoids the creation of big
arrays just to re-join and throw them away.