Support immediately require calls
Add support to modules that instant executing a required module.

    require('./something')()

This mock-fn-as-callback-style was used in test found in npm/cli:

    t.test('mock as callback style', t => {
      t.mock('../my-module.js', {
        '../sub-module.js': arg => {
          t.equal(arg, 'expected')
          t.end()
        }
      })
    })

Also added more varied situations to stress test the module replacement,
such as using t.mock within one of the defined mocks, require calls at
execution time along with tests for the mock as callback style.