wip: load TypeScript in "type":"module" context
No tests for this yet (in fact it breaks a bunch of them), but getting
close to making this proof of concept work properly:
https://github.com/tapjs/ts-esm-testing/tree/main

1. When in "type":"script" mode, files *must* be loaded with their full
   extension.  `import * from '../index'` isn't going to work.
2. That would be fine and normal and not at all surprising, *except* in
   order to get that to work, you have to `import * from 'index.js'`
   (even though `index.js` doesn't exist, and the file is _actually_
   `index.ts`).

So, caveat for those venturing into this place, you have to load the
system under test by its ".js" name.

Thankfully, we can make TypeScript do this dance, if we add
`"module":"ES2020"` to the compiler options config.

Not so thankfully, if we add `"module":"ES2020"` to the TypeScript
compiler options config, it _must_ be a `"type":"module"` package, or
else it refuses to compile.

So, ok, check the package.json, add the magic config.

Also, we have to run the file as `node ts-node-esm <filename>`.  The
`--require` hook option was fully broken.  Good, good, fine and normal.

The challenge at this point is that NYC has been rendered fully
incapable by Node's ESM changes.  So, the only way to get coverage for
projects using `"type":"module"` is to finally switch over to using C8.

Fine, that's been on the agenda forever, now's as good a time as any.
The extra work here is:

1. Figure out how to properly limit the C8 tracking to *only* the files
   specified in the coverage-map or at least *only* the files that look
   like our system under test.  It puts a LOT of data in those temp
   files, and most of it is entirely useless.

2. Probably going to have to roll our own equivalent (or add to C8 the
   equivalent) of istanbul-lib-processinfo based on C8 temp folder data,
   or else `--watch` and `--changed` won't work anymore.  C8 puts the
   pid in the filename, but I don't see any place where it's stashing
   the program or arguments, and so of course there's no externalIds
   equivalent, which is how tap knows which coverage was provided by
   which test process.

Re: https://github.com/tapjs/node-tap/issues/807