Unravel some type inferences that were failing
TS seems to get confused in some cases when a package.json-exported
submodule is called both from a CJS and ESM build.

I haven't worked out why this caused a problem for the spawn and stdin
plugins, and not the others in core/plugins, but I wouldn't be surprised
if at some point those have to be moved out as well. There doesn't seem
to be any issue pulling in modules that just export a single module,
so that is a perfectly fine workaround.