move all core plugins out, reduce ts garbage
When a class or interface is inferred from another module, as from a
return type, and that class was not exported, *or* any time a class is
re-exported from a submodule like `'@tapjs/core/plugin/whatever'`
(rather than `'@tapjs/whatever'`), TypeScript adds a property like
`__@12#private: any` to prevent it from being considered the "same" as
the original interface or used in cases where that original type is
expected.

I'm sure there's a good reason for this, but I don't like it. Those
properties don't actually exist, and in practice serve mainly to make
typedocs and autocomplete more noisy and confusing.

Thankfully, the workaround is simple, and as is usually the case when
TypeScript is being an annoying little snot, I think it's a net
improvement to the code.

All plugins now live in their own top-level workspace folder under `src`
and will be published as separate packages, and their return types are
exported, so that TS can declare them by reference directly.

In addition to making things more consistent, this also removes those
generated garbage properties, and I don't have to waste brain cycles
wondering if this or that plugin should be in core or its own package.

Exporting more types also makes the typedoc output a bit more useful
anyway. So net wins all around.