reporter: print test name more usefully on throws
If we throw after a test ends (eg, by making assertions
in excess of the plan, or after the test is ended), then
a 'test' diag is attached.

Just show that as the prefix of the result, like we would if
it had happened within the test, and strip the diag test name.