Faster build and easier bootstrapping with nx
Hopefully, this makes it possible to actually build this project in CI
and on smaller systems. I was starting to go a bit crazy trying to
manually comb through the rats nest of dep cycles here to be able to
build this thing properly if it's not already been built previously.

All sub-dependencies have been hoisted to the top level, to prevent
spurious cyclical dependency detection by nx when it builds the graph.
This also means that all projects MUST use the same versions of all
their build/lint/test tools, whether that's a bug or feature is TBD, but
I think it's nice so far.

Missing dependencies from @tapjs/run -> spawn, stdin, and reporter were
added.

@tapjs/test now uses the 'build' script instead of 'prepare', to avoid
the dependency cycle on the 'prepare' target (since its initial default
artifacts are checked in, this is a spurious cycle).