Refactor Test build, add t.applyPlugin method
In attempting to get the @tapjs/nock tests to pass, I found it's nearly
impossible with the previous setup to actually exercise the plugin
without already having it enabled. Which means, in order to test
anything here, _everything_ has to be switched on, which I definitely
don't want.

The t.applyPlugin method is definitely a bit of a hack, but it works.
From TypeScript's point of view, it only applies to that test and all
its children, and only on the return value rather than the object itself
(even though the return value is the same object). In practice, applying
any plugin anywhere applies it to all Test objects instantiated after
that point, because it pushes it into the set of plugins in use.

It'd be worthwhile to explore keeping the set of plugins on the Test
object itself somewhere, and then only using its set in child tests.
However, we don't pass any arguments to the child Test objects when they
are instantiated, because they're built by the TestBase class which owns
the flow control and subtest management logic, so this is extremely
non-trivial.

Strongly considering adding a guard to _only_ allow adding a plugin on
the root test, and _only_ if it has not yet had any child tests, which
would significantly reduce the potential for confusion about what is
available where. Ie, if you want a different plugin for one test, add it
at the top, and it's there for that file. While somewhat less
powerful/convenient, it would also sidestep any issues where a plugin is
added to one buried subtest, and surprisingly shows up for all
subsequent tests.

Apart from this feature addition, the refactoring puts much more of the
type handling logic in the built Test class into TypeScript itself, so
that the codegen can be much simpler, essentially just dumping the list
of imports, the configs, loaders, and signature.