Add skipLibCheck to run, reporter
Even though they're not in the dependency cycle, they depend on
@tapjs/test, which is pre-built. As a result, the referenced types
won't exist during a cold-start build.

The other alternative would be for these packages to depend on all of
the default build plugins, so that they don't attempt to build until
their types are present, but that would be more make work to manage, and
make the build slower with little benefit. If there's an inconsistency
in the types, we'll find it other ways easily enough.