Make t.applyPlugin() more powerful
With this change, any test can safely call `t.applyPlugin()`, as long as
it has not yet produced any output.

The resulting return value will be a different proxy, but using the
original object as its base, but *without* modifying the nature of any
other Test objects created outside its hierarchy. Any child tests that
it produces *from that point forward*, however, will have the extension.

The extended and origin objects will share the same TestBase root, so
any tests run on the extended object will be shared with the origin,
meaning that `const t = tap.applyPlugin(myPlugin)` at the top of the
test will result in a `t` object that can still be used as the root TAP
object.

This is really only relevant for testing plugins and the Test object
itself, but could become relevant if we want to be able to do
nested/extended configs, which may wish to apply different plugins by
config for different sets of tests.

This also moves the `t.plugins` getter and `t.pluginLoaded()` methods
into the `applyPlugins` extension set, and stubbed with empty returns on
the base Test class, which saves having to maintain consistency with
"the plugin set used for child tests" and "the plugins actually applied
on this object".

One weird thing is that the `.t` member on the underlying TestBase
object will now point at the newly extended Test proxy, but I think
that's an acceptable compromise.  After all, it *is* the same underlying
TestBase, just extended multiple times, and we have to choose one, so
may as well be the one that's maximally decorated.