build: place pj files in tmp, sync to dist
When running tests across all projects in parallel with nx, there is a
race condition that happens ever so often, because a package might not
have its `package.json` files placed in its dist folder at the exact
moment the test runs, resulting in a spurious error.

Also, it's just somewhat less efficient to be deleting and re-creating
those files on each build. Better to make the dist-tmp be exactly what
we want, and then sync the changes.