config: less aggressive parallelism heuristic
Getting some really annoying behaviors when running a large number of
tests on a system with a lot of CPUs, because the total processing power
gets tied up, and timers go all sideways. Plus there's not enough
processing power available to update the display, so it gets flickery,
even on fancy GPU-enabled terminals. Can't have that.

Better to leave a few available so the display is a bit smoother and
things run more consistently.