do not show full coverage text reports by default
The overlapping cases get a bit convoluted here, but I think it strikes
the right balance between noise and expectations.

By default, when coverage is full, a text report will not be shown on
test runs. It will still be shown by default on explicitly running
'tap report', unless the config is explicitly disabled.

If the config is explicitly disabled and 'tap report text' is run, then
a text summary will be printed instead of the full text report.

If the entire test suite does *not* have full coverage, then only the
files lacking full coverage will be printed, which is the useful
information in that case.

If people use it and find this behavior annoying or hard to reason
about, it's easy enough to change, but this seems like what I usually
want in most cases, and it's configurable enough to probably be fine if
the situation differs.