handle fully internal stacks less confusingly
Because all of the stack frames were omitted, the extraFromError
function and TestBase class ended up in a state where it assumed that
the stack was cleaned, but in fact it had not been modified.

Then, it tries to parse the error *message* as a stack trace call site
line, resulting in a very confusing 'at' value.

If the first line of the stack is the default error name: message, and
the parsed stack is empty, then skip it, and set at to null.