improve tests for more scenarios
There were a few inconsistencies between node 16 and node 18. When
findSourceMap gets a non-Error object, node 16 fails, whereas node 18 is
fine with it. So, only pass the error object if it's an instanceof
Error.

Also, node 16 puts a 'Function.all' callsite in several of the stacks
that were being tested, so filter that out of the snapshots.

Lastly, handle some cases where a -b argument in tests here would change
the Test output that we're testing, so that the snapshots are
consistent.