make test-built package a link dep
Having it as a bundled dep is too tricky and full of peril. Either it's
an invalid package name, and thus outside of npm's purview entirely
(meaning it won't be bundled), or it's a valid name, and npm tries to
optimize the tree by hoisting the version from the registry into the
root node_modules folder. Either is no good.

We do need *some* version of a built Test class to be present on initial
install of @tapjs/test. The downside of doing it this way is that we
can't update the Test class without also clobbering the built copy. But,
(a) that's not such a huge deal, since tap will just rebuild on the
next test run if the plugins config doesn't match what's been built, and
(b) it probably will need to be rebuilt anyway in most cases, since a
new version of @tapjs/test probably means either Test or core changed in
some way, even if only in exported types.