convert all ws to esm, improve 'threw()' behavior
It's been getting really annoying switching back and forth between CJS
and ESM when writing tests. Better to just be ESM everywhere by default,
and have the tsconfig/cjs.json configs tell tsc how to build for CJS.

When tests *must* be CommonJS, write the test with a .cjs extension
rather than using typescript for it. It's easy enough to do, since those
modules are typically tiny little shims for stuff like require.resolve
or __dirname anyway.

In the process of this refactor, shook out a bit of a weird issue with
how throws are handled by the Spawn and Worker classes. Namely, they are
effectively *always* proxied to the parent, or just ignored, because the
classes don't have any way to print TAP output. The result was that the
thrown error is reported on the parent (or, not at all), when it would
usually be more appropriate inside of the child test block in the TAP
output.

While doing this, also abstracted out the logic for getting the error
message, since that was repeated in both Base and TestBase. It's still
not quite as clean as I'd really like, but it's a big improvement.

Lots of tests had to change, of course, since they were all written with
the assumption of compiling to CJS. Some were written in CJS rather than
TypeScript, in some cases even before the advent of const/let, so some
modernization was also in order.