add t.error(), errorOrigin. refactor SuiteSummary
This restores t.error(), which was overlooked in the asserts plugin.

Print the errorOrigin callsite and stack, as well as the assert callsite
location. This settles a long-standing question about what is best to
print for a failed t.error() call. The error origin is given the same
fancy source printing treatment, along with the failed assertion
callsite.

This also makes it possible to provide the same kind of information on
`t.doesNotThrow()`, `t.resolves()`, and `t.resolveMatch()` when the
promise rejects. This commit applies that by using `this.error()` in
place of `this.#t.fail()` in the catch conditions in those assertions,
and cleans up some just clearly wrong return value logic that was in
contradiction to the documentation.

Also, refactor the SuiteSummary tag, to DRY up some of the copied code
between the terse and base report types. I'm not sure why, but this also
prevents ink from printing a bunch of blank lines sometimes at the end
of the test run, which I suspect is somehow related to using hooks in
the top level app which try to fire right at the very end as the app
unloads.