allow plugins to add testFileExtensions
The `include` option now can include the token `__EXTENSIONS__`, which
is expanded to the union set of all `testFileExtensions` string[] values
exported by loaded plugins.

This means that if you disable the typescript plugin, tap won't try to
load ts files, since they'll fail anyway.

It also means that one could conceivably add a coffeescript plugin, and
add `'coffee'` to the list, or any other JS dialect with a custom loader
and file type that one might want to add support for.