reporter: refactor exports, useLayoutEffect instead of useEffect
The reason for the flaky tests is that useEffect is deferred if called
during a render. While this is sensible on web browsers, it's quite
annoying for terminal programs, and no amount of sleeping can guarantee
that it'll actually be ready to check the output at any given time.

useLayoutEffect is the fix.

Also, export all of the hooks, utils, and components, and add a
'addType' method so that it's more easy to create custom reports.