core: clean up #end and ondone handling
This fixes a 'test incomplete' error that was occurring due to how the
'#end' was being pushed onto the queue when a promise resolves, even
though there may end up being more assertions coming as a result of
running child tests.

With this change, the Promise ending sets an '#awaitingEnd' flag. If the
test runs through its queue, and is awaiting an end, then the end will
be called with the appropriate implicit state.

Also addresses a mistake in how ondone was being handled, which was
copied from tap v16 where it made sense, given the different
architecture, but led to AsyncHook contexts never being fully cleaned
up, just exited. Not a huge deal, since there's a relatively small
number of them, but if you did have thousands of tests in a single file,
doing lots of async activities, it would potentially result in pretty
big unnecessary memory overhead.