snapshot: support CRLF line endings properly
There is really on way to do this that doesn't just fall back to binary
encoding. For example, escaping the \r as \\r means that any string
ending in "\\r\n" will fail to match. But at least it can be done
transparently, and lean on tcompare's sensible printing of Buffer
objects.

Fix: https://github.com/tapjs/tapjs/issues/641